base/transcription/whisper_cpp/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv